Fruit_Classifier/requirements.txt
# Core ML/Data Science
tensorflow==2.16.1
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.2.2

# Image Processing
Pillow==10.3.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Web Framework
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Production Server
gunicorn==21.2.0

# Load Testing
locust==2.31.2

# Jupyter/Development
jupyter==1.1.1
ipython==8.17.2